backend/requirements.txt
Django>=4.2,<5.2
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
psycopg2-binary>=2.9.0
python-decouple>=3.8
djangorestframework-simplejwt>=5.3.0
Pillow>=10.0.0

# Celery and Redis
celery>=5.3.0
redis>=5.0.0
django-celery-beat>=2.5.0
django-celery-results>=2.4.0

# Django Channels for WebSockets
channels>=4.0.0
channels-redis>=4.0.0
daphne>=4.0.0

# Django AllAuth for authentication
django-allauth>=0.60.0

# Rate limiting
django-ratelimit>=4.0.0

# Map APIs
requests>=2.31.0

# Email
sendgrid>=6.10.0

# Monitoring and logging
sentry-sdk>=2.0.0

# Testing
pytest>=8.0.0
pytest-django>=4.5.0
pytest-cov>=4.0.0

# PDF and Excel generation
reportlab>=4.0.0
openpyxl>=3.1.0

# Linting
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
